requirements.txt
Django==1.11
psycopg2==2.7.3.2

id-phonenumbers==0.3.0
phonenumbers==8.8.10
django-model-utils==3.0.0
Pillow==4.3.0